feat(asr): 桌宠 mic 按钮移入角色工具栏（截图按钮下方）+ 输入框随 AI 回复自动清空
