+++ src/components/pet/ChatInput.vue
@@ -19,15 +19,6 @@
         @compositionstart="isCompsing = true"
         @compositionend="isCompsing =false"
       />
-      <!-- 语音输入（与桌面 GameDialog 同源：useAsrInput 共享会话） -->
-      <button
-        class="h-6 px-2 bg-white/10 hover:bg-white/20 text-white rounded-full transition-colors flex items-center gap-1 shrink-0 disabled:opacity-40 disabled:cursor-not-allowed"
-        :disabled="!canStartMic"
-        :title="micTitle"
-        @click="toggleRecording"
-      >
-        <component :is="micIcon" class="w-3.5 h-3.5" />
-      </button>
       <button
         class="h-6 px-2 bg-linear-to-tr from-cyan-500 to-blue-400 hover:from-cyan-400 hover:to-blue-300 text-white font-bold text-sm rounded-full shadow-[0_4px_15px_rgba(6,182,212,0.4)] hover:shadow-[0_6px_20px_rgba(6,182,212,0.6)] transition-all duration-300 active:scale-95 flex items-center gap-1 overflow-hidden relative"
         @click="sendMessage"
@@ -50,11 +41,10 @@ import { useGameStore } from '@/stores/modules/game'
 import { useUIStore } from '@/stores/modules/ui/ui'
 import { useSettingsStore } from '@/stores/modules/settings'
 import { useLlmProvidersStore } from '@/stores/modules/llm-providers'
-import { useAsrStore } from '@/stores/modules/settings/asr'
 import { useAsrInput, registerAsrInputBridge } from '@/composables/useAsrInput'
 import { useScreenshot } from '@/composables/useScreenshot'
 import { setInputHasText } from '@/composables/useCanDeliver'
-import { Forward, Mic, MicOff } from 'lucide-vue-next'
+import { Forward } from 'lucide-vue-next'
 
 const { t } = useI18n()
 const gameStore = useGameStore()
@@ -69,47 +59,6 @@ const {
   clear: clearScreenshot,
 } = useScreenshot()
 
-// ── 语音输入（与桌面 GameDialog 同源：useAsrInput 模块级单例共享会话） ──
-const asrInput = useAsrInput()
-const asrStore = useAsrStore()
-
-// 三层状态（与 GameDialog 一致）：auto_listen 模式开 → mic = 功能开关
-const autoListenOn = computed(() => asrStore.settings.auto_listen)
-const autoListenActive = computed(() => asrInput.autoListenActive.value)
-const micIcon = computed(() => {
-  if (autoListenOn.value) return autoListenActive.value ? MicOff : Mic
-  return Mic
-})
-const micTitle = computed(() => {
-  if (autoListenOn.value) {
-    return autoListenActive.value
-      ? t('game.dialog.asrAutoOff') // 监听中：暂停
-      : t('game.dialog.asrAutoResume') // 已暂停：恢复
-  }
-  return asrInput.phase.value === 'recording'
-    ? t('game.dialog.recordingStop')
-    : t('game.dialog.voiceInput')
-})
-const canStartMic = computed(
-  () =>
-    asrStore.settings.voice_input_enabled &&
-    (autoListenOn.value || asrInput.phase.value === 'recording' || asrInput.canStartAsr()),
-)
-function toggleRecording() {
-  // auto_listen 模式开：mic 按钮 = 切换功能开关（暂停/恢复监听），不改模式设置
-  if (autoListenOn.value) {
-    asrInput.toggleAutoListenFunction()
-    return
-  }
-  if (asrInput.phase.value === 'idle') {
-    void asrInput.start('button').catch(() => {
-      /* 会话忙时静默忽略 */
-    })
-  } else if (asrInput.phase.value === 'recording') {
-    asrInput.stop()
-  }
-}
-
 // fill_only：识别完成整句填入（与桌面 onAsrText 一致）
 function onAsrText(e: Event) {
   const ce = e as CustomEvent<string>
@@ -118,9 +67,11 @@ function onAsrText(e: Event) {
   }
 }
 
+// 输入桥：流式 partial 实时写入（与桌面 GameDialog 一致；录音发起窗口的
+// phase 是窗口本地状态，partial 只写入发起方输入框）
+const asrInput = useAsrInput()
 onMounted(() => {
   initScreenshot()
-  // 输入桥：流式 partial 实时写入（只在录音发起窗口生效——phase 是窗口本地状态）
   registerAsrInputBridge({
     getText: () => messageText.value,
     setText: (v) => {
@@ -132,6 +83,14 @@ onMounted(() => {
 onUnmounted(() => {
   window.removeEventListener('asr-text', onAsrText)
   destroyScreenshot()
+})
+
+// 与主对话窗口行为一致（GameDialog 同款 watch）：AI 回复（showCharacterLine
+// 非空 + responding）时清空输入框——auto_send 识别文本填入后随回复自动清空
+watch([() => uiStore.showCharacterLine, () => gameStore.currentStatus], ([newLine, newStatus]) => {
+  if (newLine && newLine !== '' && newStatus === 'responding') {
+    messageText.value = ''
+  }
 })
 
 const scale = computed(() => settingsStore.pet?.scale || 1.0)
